CURD operations functional

+++ utils/types.go
@@ -4,23 +4,22 @@ import "fmt"
 
 type Inventory struct {
 	name string
-	cost float32
+	cost float64
 	desc string
 }
 
 type City struct {
 	name string
-	quan int
+	quan int64
 	city string
 }
 
 func (item Inventory) print() {
 	fmt.Printf("name:%s\nprice:%f\ndescription:%s\n",
-							item.name, item.cost, item.desc)
+		item.name, item.cost, item.desc)
 }
 
 func (item City) print() {
 	fmt.Printf("name:%s\nquantity:%d\ncity:%s\n",
-							item.name, item.quan, item.city)
+		item.name, item.quan, item.city)
 }
-
